bug fix ndcHit always return (-1, -1, z)
(Internal change: 2003955)

+++ pxr/imaging/lib/hdx/pickTask.cpp
@@ -636,8 +636,8 @@ HdxPickResult::_ResolveHit(int index, int x, int y, float z,
 
     // Calculate the hit location in NDC, then transform to worldspace.
     GfVec3d ndcHit(
-        (x / _bufferSize[0]) * 2.0 - 1.0,
-        (y / _bufferSize[1]) * 2.0 - 1.0,
+        ((double)x / _bufferSize[0]) * 2.0 - 1.0,
+        ((double)y / _bufferSize[1]) * 2.0 - 1.0,
         ((z - _depthRange[0]) / (_depthRange[1] - _depthRange[0])) * 2.0 - 1.0);
     hit->worldSpaceHitPoint = GfVec3f(_ndcToWorld.Transform(ndcHit));
     hit->ndcDepth = ndcHit[2];
